Add architecture documentation and refactor accordingly
