+++ src/main/java/com/philips/research/spdxbuilder/core/ConversionStore.java
@@ -10,8 +10,8 @@
 
 package com.philips.research.spdxbuilder.core;
 
-import com.philips.research.spdxbuilder.core.bom.BillOfMaterials;
-import com.philips.research.spdxbuilder.core.bom.Package;
+import com.philips.research.spdxbuilder.core.domain.BillOfMaterials;
+import com.philips.research.spdxbuilder.core.domain.Package;
 
 import java.io.File;
 import java.net.URI;
