+++ src/main/java/com/philips/research/spdxbuilder/core/domain/BillOfMaterials.java
@@ -8,7 +8,7 @@
  * All Rights Reserved
  */
 
-package com.philips.research.spdxbuilder.core.bom;
+package com.philips.research.spdxbuilder.core.domain;
 
 import pl.tlinkowski.annotation.basic.NullOr;
