+++ src/main/java/com/philips/research/spdxbuilder/core/domain/Party.java
@@ -8,7 +8,7 @@
  * All Rights Reserved
  */
 
-package com.philips.research.spdxbuilder.core.bom;
+package com.philips.research.spdxbuilder.core.domain;
 
 /**
  * Named person, tool or organization.
