+++ src/test/java/com/philips/research/spdxbuilder/core/domain/BillOfMaterialsTest.java
@@ -8,7 +8,7 @@
  * All Rights Reserved
  */
 
-package com.philips.research.spdxbuilder.core.bom;
+package com.philips.research.spdxbuilder.core.domain;
 
 import org.junit.jupiter.api.Test;
