+++ src/test/java/com/philips/research/spdxbuilder/persistence/ConversionPersistenceTest.java
@@ -11,7 +11,7 @@
 package com.philips.research.spdxbuilder.persistence;
 
 import com.philips.research.spdxbuilder.core.ConversionStore;
-import com.philips.research.spdxbuilder.core.bom.Package;
+import com.philips.research.spdxbuilder.core.domain.Package;
 import com.philips.research.spdxbuilder.persistence.license.LicenseScannerClient;
 import org.junit.jupiter.api.Test;
